ФЗ-11

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -34,7 +34,7 @@ public class FilmService {
                 .stream()
                 // сортируем по убыванию количества лайков
                 .sorted((one, two) -> Integer.compare(two.getWhoLikes().size(), one.getWhoLikes().size()))
-                .limit((long) quantity)
+                .limit(quantity)
                 .collect(Collectors.toList());
     }
 }
